+++ README.md
@@ -2,6 +2,8 @@
 
 wanted preonboarding challenge
 
+# [API ReadME](https://github.com/starkoora/wanted-pre-onboarding-challenge-fe-1-api)
+
 # 프로젝트 셋팅
 
 1. yarn 버전을 berry 로 변경 및 프로젝트 초기화
